Move messages to their own dir, add mocha & chai
