+++ typescript/sdk/.mocharc.json
@@ -1,0 +1,3 @@
+{
+  "require": ["ts-node/register"]
+}
