+++ typescript/sdk/package.json
@@ -8,18 +8,21 @@
     "build": "tsc",
     "check": "tsc --noEmit",
     "prettier": "prettier --write ./src",
-    "prepublishOnly": "npm run build"
+    "prepublishOnly": "npm run build",
+    "test": "mocha --config .mocharc.json ./test/**/*.test.ts"
   },
   "license": "MIT OR Apache-2.0",
   "devDependencies": {
     "@types/node": "^16.9.1",
     "@typescript-eslint/eslint-plugin": "^4.33.0",
     "@typescript-eslint/parser": "^4.33.0",
+    "chai": "^4.3.6",
     "dotenv": "^10.0.0",
     "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.3.0",
     "fs": "0.0.1-security",
     "prettier": "^2.4.1",
+    "mocha": "^9.2.2",
     "typescript": "^4.4.3"
   },
   "dependencies": {
