+++ typescript/sdk/src/core/message/undispatched-message.ts
@@ -1,12 +1,9 @@
 import { arrayify, hexlify } from '@ethersproject/bytes';
-// import { ethers } from 'ethers';
 import { keccak256 } from 'ethers/lib/utils';
 
 import { Inbox, Outbox } from '@abacus-network/core';
 
-// import { NameOrDomain } from '../types';
-
-import { AbacusCore } from '.';
+import { AbacusCore } from '..';
 
 export type ParsedMessage = {
   from: number;
@@ -53,59 +50,6 @@ export class UndispatchedMessage {
     this.outbox = core.mustGetContracts(this.message.from).outbox;
     this.inbox = core.mustGetInbox(this.message.from, this.message.destination);
   }
-
-  // /**
-  //  * Estimates the amount of gas required to process a message.
-  //  * This does not assume the Inbox of the domain the message will be processed on has
-  //  * a checkpoint that the message is included in. Therefore, we estimate
-  //  * the gas by summing:
-  //  * 1. The intrinsic gas cost of a transaction on the destination chain.
-  //  * 2. Any gas costs imposed by operations in the Inbox, including proving
-  //  *    the message and logic surrounding the processing of a message.
-  //  * 3. Estimating the gas consumption of a direct call to the `handle`
-  //  *    function of the recipient address using the correct parameters and
-  //  *    setting the `from` address of the transaction to the address of the inbox.
-  //  * 4. A buffer to account for inaccuracies in the above estimations.
-  //  */
-  // async estimateGas(
-  //   src: NameOrDomain,
-  //   dst: NameOrDomain,
-  //   sender: string,
-  //   recipient: string,
-  //   message: string,
-  // ) {
-  //   // TODO come back to this
-  //   const intrinsicGas = 21_000;
-  //   const provingAndProcessingInboxCosts = 120_000;
-
-  //   const provider = this.core.mustGetProvider(dst);
-  //   const inbox = this.core.mustGetInbox(src, dst);
-
-  //   const srcDomain = this.core.mustGetDomain(src);
-
-  //   const handlerInterface = new ethers.utils.Interface([
-  //     'function handle(uint32,bytes32,bytes)',
-  //   ]);
-  //   const directHandleEstimation = await provider.estimateGas({
-  //     to: recipient,
-  //     from: inbox.address,
-  //     data: handlerInterface.encodeFunctionData('handle', [
-  //       srcDomain.id,
-  //       sender,
-  //       message,
-  //     ]),
-  //   });
-
-  //   console.log(
-  //     'directHandleEstimation',
-  //     directHandleEstimation,
-  //     directHandleEstimation.toNumber(),
-  //   );
-
-  //   return directHandleEstimation
-  //     .add(intrinsicGas)
-  //     .add(provingAndProcessingInboxCosts);
-  // }
 
   /**
    * The domain from which the message was sent
