+++ typescript/sdk/src/types.ts
@@ -46,6 +46,7 @@ export type ChainName = MainnetChainNames | TestnetChainNames | TestChainNames;
 export interface Domain {
   id: number;
   name: ChainName;
+  nativeTokenDecimals?: number,
   paginate?: Pagination;
 }
