+++ typescript/sdk/test/core/message/interchain-gas.test.ts
@@ -1,0 +1,7 @@
+import { expect } from 'chai';
+
+describe('InterchainGasPayingMessage', () => {
+  it('can run a test', () => {
+    expect(true).to.be.true;
+  });
+});
