+++ typescript/sdk/tsconfig.json
@@ -4,7 +4,7 @@
     "outDir": "./dist/",
     "rootDir": "./src/"
   },
-  "exclude": ["./node_modules/", "./dist/"],
+  "exclude": ["./node_modules/", "./dist/", "./test/"],
   "include": [
     "./src/*.ts",
     "./src/**/*.ts"
